Troubleshooting zoxide, nushell, and hyprland. bool to string
I thought there was an issue with the way nushell was treating the
autoload directory for my initialization of the zoxide utility. The
environment variable for the zoxide hook read as
"__zoxide_hooked"="true" vs. __zoxide_hooked=true, with the latter being
the intended expression.

When troubleshooting in a tty, `$env.__zoxide_hooked | describe` outputs
`bool`, while after launching Hyprland the same command outputs
`string`. This makes me believe the offender was Hyprland all along.
Although Hyprland is not turning all of my environment variables into
string, but this one it seems like it is?

Anyways, I fixed the issue by manually adding this to my config.nu file:
`$env.__zoxide_hooked = default false`, as this seems the intended way
from the auto-generated script.