Fixing mail setup.
This was hard work. I tried to get nushell to cooperate but not sure if
there's a solution to get the bash equivalent of grabbing the arguments
passed in with nushell. If there is one, t hen I don't know of it. I was
very close to a solution, but until I find the answer, this mailto.nu
script (which is very wrong, even though I got to "almost" a perfect
answer) will stay as it is, since there is no reason harming myself any
longer to find a solution. :(